requirements.txt
pandas
numpy
scikit-learn
flask
seaborn
matplotlib
joblib
json
request
jsonify
path
time